backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
alembic==1.13.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
cryptography==43.0.1
apscheduler==3.10.4
python-multipart==0.0.10
httpx==0.27.2
